small change on follow request decline

+++ InTouch-Backend/Data/Services/UsersService.cs
@@ -238,11 +238,15 @@ namespace InTouch_Backend.Data.Services
             var user = _context.Users.FirstOrDefault(u => u.Id == id);
             if (user != null)
             {
-                string ImageFilePath = Path.Combine(Directory.GetCurrentDirectory(), "User Images", user.ImagePath);
-                if (File.Exists(ImageFilePath))
+                if (!string.IsNullOrEmpty(user.ImagePath))
                 {
-                    File.Delete(ImageFilePath);
+                    string ImageFilePath = Path.Combine(Directory.GetCurrentDirectory(), "User Images", user.ImagePath);
+                    if (File.Exists(ImageFilePath))
+                    {
+                        File.Delete(ImageFilePath);
+                    }
                 }
+
                 var userPosts = _context.Posts.Where(u => u.userID == id);
                 foreach (var item in userPosts)
                 {
@@ -265,7 +269,6 @@ namespace InTouch_Backend.Data.Services
 
                 var userSavedPosts = _context.SavedPosts.Where(s => s.UserId == id);
                 _context.SavedPosts.RemoveRange(userSavedPosts);
-
